requirements.txt
numpy==1.24.3
pandas==2.0.3
matplotlib==3.7.2
seaborn==0.12.2
opencv-python==4.8.0.76
Pillow==9.5.0
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.1
pyyaml==6.0.1
scipy==1.11.2
streamlit==1.26.0
plotly==5.16.1
